Put platform-specific code into subdirs
All the windows specific code is now in the windows subdir. The goal of
this patch is to preserve druid-shell function on windows, and have it
compile (with basically no functionality) on mac.

Part of #13
