+++ druid-shell/Cargo.toml
@@ -3,7 +3,7 @@ name = "druid-shell"
 version = "0.1.1"
 license = "Apache-2.0"
 authors = ["Raph Levien <raph.levien@gmail.com>"]
-description = "Windows-specific application shell used for druid toolkit."
+description = "Platform abstracting application shell used for druid toolkit."
 repository = "https://github.com/xi-editor/druid"
 categories = ["os::windows-apis", "gui"]
 
@@ -11,15 +11,17 @@ categories = ["os::windows-apis", "gui"]
 default-target = "x86_64-pc-windows-msvc"
 
 [dependencies]
-directwrite = "0.1.2"
-direct2d = "0.2.0"
-wio = "0.2"
 piet-common = "0.0.1"
 
 lazy_static = "1.0"
 time = "0.1.39"
 
-[dependencies.winapi]
+[target.'cfg(target_os="windows")'.dependencies]
+directwrite = "0.1.2"
+direct2d = "0.2.0"
+wio = "0.2"
+
+[target.'cfg(target_os="windows")'.dependencies.winapi]
 version = "0.3.6"
 features = ["d2d1_1", "dwrite", "winbase", "libloaderapi", "errhandlingapi", "winuser", "shellscalingapi", "shobjidl", "combaseapi", "synchapi", "dxgi1_3", "dcomp", "d3d11", "dwmapi", "wincon", "fileapi", "processenv", "winbase", "handleapi"]
