+++ druid-shell/examples/hello.rs
@@ -27,7 +27,7 @@ use druid_shell::dialog::{FileDialogOptions, FileDialogType};
 use druid_shell::menu::Menu;
 use druid_shell::win_main;
 use druid_shell::window::{MouseEvent, WinHandler, WindowHandle};
-use druid_shell::windows::WindowBuilder;
+use druid_shell::platform::WindowBuilder;
 
 #[derive(Default)]
 struct HelloState {
