+++ druid-shell/src/windows/dialog.rs
@@ -28,7 +28,8 @@ use wio::com::ComPtr;
 
 use std::ffi::OsString;
 use std::ptr::null_mut;
-use util::{as_result, Error, FromWide};
+use util::{as_result, FromWide};
+use Error;
 
 /// Type of file dialog.
 pub enum FileDialogType {
