+++ druid-shell/src/windows/paint.rs
