+++ examples/anim.rs
@@ -23,7 +23,7 @@ use kurbo::Line;
 use piet::RenderContext;
 
 use druid_shell::win_main;
-use druid_shell::windows::WindowBuilder;
+use druid_shell::platform::WindowBuilder;
 
 use druid::{Ui, UiMain, UiState};
