+++ examples/calc.rs
@@ -18,7 +18,7 @@ extern crate druid;
 extern crate druid_shell;
 
 use druid_shell::win_main;
-use druid_shell::windows::WindowBuilder;
+use druid_shell::platform::WindowBuilder;
 
 use druid::widget::{Button, Column, EventForwarder, KeyListener, Label, Padding, Row};
 use druid::{KeyEvent, KeyVariant, UiMain, UiState};
