+++ examples/dynamic.rs
@@ -20,7 +20,7 @@ extern crate druid_shell;
 use std::collections::BTreeMap;
 
 use druid_shell::win_main;
-use druid_shell::windows::WindowBuilder;
+use druid_shell::platform::WindowBuilder;
 
 use druid::widget::{Button, Column, EventForwarder, Label, Padding, Row};
 use druid::{Id, UiMain, UiState};
