+++ examples/ext_event.rs
@@ -20,7 +20,7 @@ extern crate druid_shell;
 use std::{thread, time};
 
 use druid_shell::win_main;
-use druid_shell::windows::WindowBuilder;
+use druid_shell::platform::WindowBuilder;
 
 use druid::widget::Label;
 use druid::{UiMain, UiState};
